add hook to goal board
