+++ libs/web/feature-board/src/lib/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './useGoalBoard';
